Revert "Model picker"
