+++ requirements.txt
@@ -1,7 +1,6 @@
 ctranslate2
 faster-whisper
 huggingface-hub
-opencv-python==4.10.0.84
 numpy
 pyaudio
 pyautogui
